requirements.txt
torch
torchvision
transformers
Pillow
requests
beautifulsoup4
PyPDF2
pdf2image
numpy
scipy
opencv-python
scikit-learn
scikit-image
joblib
tqdm
matplotlib
seaborn
plotly
tensorboard
wandb
albumentations
timm
efficientnet-pytorch
segmentation-models-pytorch
captum
lime
shap
gradio